Improve debug output
