+++ rollup.config.js
@@ -8,6 +8,18 @@ export default {
     format: 'cjs',
     exports: 'named',
   },
-  external: ['fs/promises', 'path', 'util'],
+  external: [
+    '@actions/core',
+    '@actions/github',
+    '@octokit/action',
+    'debug',
+    'fs/promises',
+    'graphql-request',
+    'graphql',
+    'iso8601-duration',
+    'path',
+    'pretty-ms',
+    'util',
+  ],
   plugins: [typescript(), json()],
 };
